♻️  Split up ChangesSubscriptionSender into multiple classes
This should better adhere to SRP and simplify things.
