+++ inc/Subscriptions/SubscriptionSender.php
@@ -6,7 +6,7 @@ use Mailer;
 
 abstract class SubscriptionSender
 {
-    private $mailer;
+    protected $mailer;
 
     public function __construct(Mailer $mailer = null)
     {
@@ -16,6 +16,30 @@ abstract class SubscriptionSender
         $this->mailer = $mailer;
     }
 
+    /**
+     * Get a valid message id for a certain $id and revision (or the current revision)
+     *
+     * @param string $id  The id of the page (or media file) the message id should be for
+     * @param string $rev The revision of the page, set to the current revision of the page $id if not set
+     *
+     * @return string
+     */
+    protected function getMessageID($id, $rev = null)
+    {
+        static $listid = null;
+        if (is_null($listid)) {
+            $server = parse_url(DOKU_URL, PHP_URL_HOST);
+            $listid = join('.', array_reverse(explode('/', DOKU_BASE))) . $server;
+            $listid = urlencode($listid);
+            $listid = strtolower(trim($listid, '.'));
+        }
+
+        if (is_null($rev)) {
+            $rev = @filemtime(wikiFN($id));
+        }
+
+        return "<$id?rev=$rev@$listid>";
+    }
 
     /**
      * Helper function for sending a mail
